transfer validator logic to seperate file
